+++ src/graphql/resolvers/comment.ts
@@ -1,8 +1,8 @@
 import { Args, FieldResolver, Query, Resolver, Root } from 'type-graphql'
 import { getAllComments, getCommentAuthor } from 'data/comments'
-import { Comment } from 'graphql/schema/Comment.schema'
-import { PaginationArgs } from 'graphql/schema/sharedArguments'
-import { User } from 'graphql/schema/User.schema'
+import { PaginationArgs } from 'graphql/schema/arguments/pagination'
+import { Comment } from 'graphql/schema/types/comment'
+import { User } from 'graphql/schema/types/user'
 
 @Resolver(Comment)
 export class CommentResolver {
