+++ src/graphql/resolvers/user.ts
@@ -1,10 +1,11 @@
 import { Arg, Args, createUnionType, Field, FieldResolver, ID, InputType, Mutation, ObjectType, Query, Resolver, Root } from 'type-graphql'
 import { getUserPosts } from 'data/posts'
 import { createUser, deleteUser, getAllUsers } from 'data/users'
-import { Post } from 'graphql/schema/Post.schema'
-import { PaginationArgs } from 'graphql/schema/sharedArguments'
-import { User } from 'graphql/schema/User.schema'
-import { ErrorCode, UserError } from 'graphql/schema/UserError.schema'
+import { PaginationArgs } from 'graphql/schema/arguments/pagination'
+import { ErrorCode } from 'graphql/schema/enums/errorCode'
+import { Post } from 'graphql/schema/types/post'
+import { User } from 'graphql/schema/types/user'
+import { UserError } from 'graphql/schema/types/userError'
 
 @ObjectType()
 export class CreateUserSuccess {
