+++ src/graphql/schema/types/comment.ts
@@ -1,5 +1,5 @@
 import { Field, ID, ObjectType } from 'type-graphql'
-import { User } from './User.schema'
+import { User } from './user'
 
 @ObjectType({ description: 'Represents a comment from a Post created by a User' })
 export class Comment {
