+++ src/graphql/schema/types/user.ts
@@ -1,5 +1,5 @@
 import { Field, ID, ObjectType } from 'type-graphql'
-import { Post } from './Post.schema'
+import { Post } from './post'
 
 @ObjectType()
 export class User {
